+++ app/Http/Controllers/JobController.php
@@ -11,14 +11,18 @@ use App\Models\Job;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreJobRequest;
 use App\Http\Requests\UpdateJobRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests; // trait für Authorize importieren 
 
 class JobController extends Controller
 {
+    use AuthorizesRequests; // ermöglicht das prüfen / authorisieren. Darf der aktuelle benutzer die Methode auf das Objekt aufrufen.
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny', Job::class);
         $jobs = Job::with(['company', 'category'])->get();
         return view('jobs.index', compact('jobs'));
     }
@@ -28,6 +32,7 @@ class JobController extends Controller
      */
     public function create()
     {
+      $this->authorize('create', Job::class);
       return view('jobs.create');
     }
 
@@ -36,6 +41,7 @@ class JobController extends Controller
      */
     public function store(StoreJobRequest $request)
     {
+        $this->authorize('create', Job::class);
         Job::create($request->validated());
         return redirect()->route('jobs.index')->with('success', 'Job wurde erstellt!'); 
     }
@@ -45,6 +51,7 @@ class JobController extends Controller
      */
     public function show(Job $job)
     {
+        $this->authorize('view', $job);
         return view('jobs.show', compact('job'));
     }
 
@@ -53,6 +60,7 @@ class JobController extends Controller
      */
     public function edit(Job $job)
     {
+        $this->authorize('update', $job);
        return view('jobs.edit', compact('job')); 
     }
 
@@ -65,6 +73,8 @@ class JobController extends Controller
         // dd($request->all()); 
 
         // Automatische Prüfung: Wenn Ablaufdatum auf die Vergangenheit gesetzt wird, Job inaktiv setzen
+        $this->authorize('update', $job);
+
         $data = $request->validated();
         if (isset($data['expires_at']) && \Carbon\Carbon::parse($data['expires_at'])->isPast()) {
         $data['is_active'] = false;
@@ -80,7 +90,20 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        $this->authorize('delete', Job::class);
         $job->delete();
         return redirect()->route('jobs.index')->with('success', 'Job wurde gelöscht!'); 
     }
 }
+
+
+/**
+ * Merksatz beispiel:
+ * $this->authorize('update', Job::class);
+ * 1. Gehe in JobPolicy
+ * 2. Rufe Methode update() auf
+ * 3. Wenn true → weiter mit nächster Zeile || Wenn false → 403 Fehler
+ * 4  true = return view('jobs.edit', compact('job')); 
+ * 5. routing zu view jobs.edit 
+ 
+ */
